feat(echo): add a dynamic toggle for full executions
Big executions can generate a lot of tasks, which means a lot of echo traffic that all include the _full_ execution payload.

I think we can definitely skip these at least for task events as a first step. We can also probably be more selective with stage events, but leaving that for later.