fix(mine): Fix canary stage cancellation
Groovy... Cancelling during [mine] canary deploys would cause an unhandled exception (`MissingMethod`) in the cancellation thread.
Because it's in a separate thread and uncaught we don't actually propagate the error in any way, but the side-effect is that if,
in a complex pipeline, some other stage fails and causes pipeline termination during a canary deploy, those deployed [canary and baseline]
clusters won't be cleaned up and still, potentially, taking traffic!

This fixes the cancelation code to correctly look up the top level canary stage and cancel it.