perf(test): Speed up PgSQL tests
When overriding setupSpec, Spock runs both the parent (super.setupSpec)
and the subclass setupSpec (in that order). This means that for the
PgSQL tests, we're creating a MySQL database then immediately discarding
it to replace it with a PgSQL database.

As creating the database spins up a container and is resonably slow,
lightly refactor this so that we don't do this extra work.