refactor(test): Push some logic down from execution repository tests
There's a lot of coupling between PipelineExecutionRepositoryTck and
its subclasses that causes initialization order dependencies. In order
to make it easier to make changes (such as changing what fields are
updated in setup vs setupSpec), refactor this so that each subclass
is fully responsible for initialization and we can control the order.

The main issue here is that the super class was storing and
initializing the ExecutionRepository, but we can't actually create
the ExecutionRepository until we've created the persistence store,
which is owned by the implementations. So we were subtly depending
on each subclass having already initialized its persistence store
before we create the ExecutionRepository.

Now it's up to each subclass ot handle this order-dependence and
expose a method to get the ExecutionRepository.