feat(kubernetes): DeployManifestStage makes live calls to determine cluster members for rollout strategies
In order to make the `liveManifestCalls` behavior the default for all users, we must change all pipeline operations that rely on cache consistency to instead call Clouddriver endpoints that make live queries against the Kubernetes cluster.

This commit replaces the existing `getClusters` call in the DeployManifestStage, used to hydrate the context of any Disable and Delete stages necessary in order to enact a Red/Black or Highlander rollout strategy, with a new `getClusterManifests` call, which returns a list of ManifestCoordinates for members of a given cluster. Also adds tests asserting on the basic correctness of the stage graph when using rollout strategies.