fix(restart): Fix parent pipeline restart for dependent pipelines
Imagine the following pipeline topology:

```
Pipeline A

Pipeline B (triggers of SUCCESS of A)
\
 \-Pipeline C (Triggered by a stage in B)
```

Now, imagine you restart a failed stage in `Pipeline C`. This stage will attempt to (correctly) restart `Pipeline B` at the appropriate stage.
This works unless `Pipeline B` was started as a result of dependent pipeline starter. In this case, restarting `Pipeline B` will fail because we will fail to restart `Pipeline A`.
This happens because while `A` is technically a parent of `B` no `parentExecutionStageId` exists so we fail to look up the stage and there fore `Pipeline B` will never actually restart as a result of restart of `C`

This change skips restarting parent if no `parentExecutionStageId` is specified