fix(wait): Fix wait for up instances task
This change in clouddriver https://github.com/spinnaker/clouddriver/pull/4681
added a boolean `pinned` field to the `Capacity` object in `ServerGroup`.
However there's some code in the `WaitForUpInstancesTask` in Orca that
depends on the types all being `Integer`. The end result is that the
wait task fails with a stack trace like:

```
org.codehaus.groovy.runtime.typehandling.GroovyCastException: Cannot cast object 'true' with class 'java.lang.Boolean' to class 'java.lang.Integer'
```

This patch updates the task to just get the fields it needs to compare
and adds a test for this case.