fix(waitforup): Don't use the min of snapshot capacity
There are situations where we capture the snapshot capacity that is wrong. For example, if we resize a cluster `1->10`, then we do a force cache refresh and if that doesn't actually get the new state of `10`,
`WaitForUp` will wait for `1` instead of `10` instances (we defer to `WaitForUp` from `WaitForCapacityMatch`).
The other situation is when we do a rollback to a `0` size ASG, the first step of resizing it will resize it to correct rollback number, but since `WaitForUp` will succeed immediately we will tear down the old ASG and thus leave a rollback with no ASG ready to take traffic.
This is mostly an issue on Titus because `ForceRefresh` is a no-op on Titus.

It appears there already was an attempt to "fixing" this issue in https://github.com/spinnaker/orca/pull/2923, however it doesn't seem to cover all the cases.

This is an initial attempt that still needs figuring out that it doesn't break any other scenarious