feat(kubernetes): identify and retry if the deploy manifest k8s task owner died while orca is monitoring the task
Before this, DeployManifestStage uses MonitorKatoTask to wait for a DeployManifestTask operation to complete. If the task isn't complete, MonitorKatoTask does have logic to call kato.resumeTask (POST task/{id}:resume, clouddriver's OperationsController.resumeTask), but that only applies when the task is retryable. Before this change only tasks in sagas are potentially retryable, so then it's up to orca to re-execute the MonitorKatoTask, and all that's happening is waiting for clouddriver to do its work.

This PR introduces a MonitorDeployManifestTask that starts a new DeployManifestTask if it detects that the clouddriver pod executing the current DeployManifestTask is no longer present. This is only relevant if clouddriver itself is running in kubernetes, and that clouddriver is configured with a kubernetes account for "itself".

Kubernetes DeployManifestTask tasks in orca correspond to KubernetesDeployManifestOperations in clouddriver. That operation is responsible for kubectl apply (or similar based on annotations in the manifest(s) being deployed). It's not typically a long operation, but if there are multiple (or large) manifests, or unreliable communications with the target cluster, it's certainly possible for the clouddriver pod to die before it completes.

The following configuration flags (and their defaults) control the behavior of this new feature:

tasks.monitor-kato-task.kubernetes.deploy-manifest:
  retry-task.enabled: false
  maximum-period-inactivity-ms: 300000L
  maximum-forced-retries: 3
  account: "" # the name of the account where clouddriver runs, required if enabled is true
  namespace: spinnaker

The execution context of kubernetes deploy manifest stages contain the following information to track the progress of additional attempts:

kato.task.forcedRetryAttempts: this is more for debugging....if forcedRetryAttempts < forcedRetries then forceRetryFatalError is true
kato.task.forcedRetries: the number of retries
kato.task.forceRetryFatalError: true if there was an error retrying

It's a little unfortunate that restarting the DeployManifestTask means creating a whole new task. Clouddriver's task repository doesn't store its inputs, at least for non-saga tasks, so it's not possible to restart an existing one. See OperationsService.collectAtomicOperationsFromSagas for more.