feat(jenkins): Stop Jenkins job when job name as slashes in the job name
When defining a Jenkins job inside folders, the name contains slashes. Because of that instead of matching the request to the IGOR STOP endpoint (/masters/{name}/jobs/{jobName}/stop/{queuedBuild}/{buildNumber}), Spring is matching the request to the BUILD one  (/masters/{name}/jobs/**)

The stop request is failing because it is trying to start a job that does not exist.

Feature flag added to call the existing endpoint (which accepts the job name as path variable) or the new one (which accepts  the job name as a query parameter)