feat(igor): Stop Jenkins job when job name has slashes in the job name (#4294)
* feat(jenkins): Stop Jenkins job when job name as slashes in the job name

When defining a Jenkins job inside folders, the name contains slashes. Because of that instead of matching the request to the IGOR STOP endpoint (/masters/{name}/jobs/{jobName}/stop/{queuedBuild}/{buildNumber}), Spring is matching the request to the BUILD one  (/masters/{name}/jobs/**)

The stop request is failing because it is trying to start a job that does not exist.

Feature flag added to call the existing endpoint (which accepts the job name as path variable) or the new one (which accepts  the job name as a query parameter)

* fix: Apply spotless fix to format the files

* refactor(igor): Implement code review comments

* refactor(review): Implement the comment from the review

* refactor(review): Implement the comment from the review

(cherry picked from commit 9e381161965e13b761e3516305b4e2abd66ff0bc)