feat(mpt-v1): Support for jinja expressions in stages
Adds support for using jinja expressions at the top level in the stages block. Because the template file had to be both well-formed YAML and adhering to the schema defined in `PipelineTemplate`, there was no way to use expressions at the top level here, as can be seen in the `dynamic-stages-001.yml` file. With this patch Orca will first load the yaml file as a `Map` instead of directly as a `PipelineTemplate`. Then it will parse any variables to create a context for the Jinja parser, and if the stages section is a single string, it will parse any Jinja expressions in the stages section. Finally it will load the result as `PipelineTemplate`, just as before, and proceed with processing the rest of the handler chain.