fix(queue): reset ackAttempts on successful SQL ack
In the SQL queue implemementation, reset ackAttempts from the
message metadata when a message is ack'd. This avoids having this
number grow over time for very long-lived messages that keep getting
re-enqueued, with the occasional failure causing a message to not be
acknowledged and eventually get dropped once 5 acks have been missed
in total.