refactor(retrofit): change BaseRetrofitExceptionHandler.shouldRetry to take an Exception instead of only RetrofitError
The logic in the implementation is still specific to retrofit as it looks for
retrofit-specific annotations in the stack trace.  This way though,
SpinnakerNetworkExceptions caused by e.g. IOExceptions (e.g. with no underlying
RetrofitError in the causal chain) are considered retryable when they happen in an
idempotent retrofit request.