refactor(deploymentmonitor): Configure SpinnakerRetrofitErrorHandler for DeploymentMonitorService
Update the error handler for the retrofit client, DeploymentMonitorService, to use SpinnakerRetrofitErrorHandler. Additionally, modify catch blocks with RetrofitError to use Spinnaker*Exception.

Purpose : This commit is foundational work for upgrading the Retrofit version to retrofit2.x

Behavioural change : The modification includes updating catch blocks, leading to a change in the format of logger messages. This aligns with the upcoming modifications discussed in the PR: https://github.com/spinnaker/orca/pull/4608

Additionally, when the error response body is null, an appropriate error message will be printed in the logger.